feat: importación masiva de jugadores con foto y carga de fotos por ZIP
- Amplía el número de camiseta de jugadores de 1-99 a 0-999 (TINYINT a
  SMALLINT en la BD) para soportar numeraciones reales fuera de ese rango.
- El CSV de importación de jugadores acepta una columna opcional "foto"
  (nombre de archivo ya existente en uploads/jugadores/) que asigna
  foto_url sin sobrescribir la foto actual cuando viene vacía.
- Nueva pantalla para subir un ZIP de fotos por equipo, donde cada
  archivo se empareja con su jugador por número de camiseta.
- Extrae la validación/guardado de imágenes de handle_image_upload() a
  guardar_imagen_validada(), reutilizada por ambos flujos.

+++ install/schema.sql
@@ -89,7 +89,7 @@ CREATE TABLE IF NOT EXISTS jugadores (
     id INT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     equipo_id INT UNSIGNED NOT NULL,
     nombre VARCHAR(100) NOT NULL,
-    numero TINYINT UNSIGNED NOT NULL,
+    numero SMALLINT UNSIGNED NOT NULL,
     posicion ENUM('Portero','Defensa','Mediocampista','Delantero') NOT NULL,
     cedula VARCHAR(20),
     foto_url VARCHAR(255),
